# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/labs": {
      "post": {
        "operationId": "LabsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      },
      "get": {
        "operationId": "LabsController_findAll",
        "parameters": [
          {
            "name": "workshopId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HandsOnLabMongo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      }
    },
    "/labs/{labId}": {
      "get": {
        "operationId": "LabsController_findOne",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      },
      "patch": {
        "operationId": "LabsController_update",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandsOnLabMongo"
                }
              }
            }
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      },
      "delete": {
        "operationId": "LabsController_remove",
        "parameters": [
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Valid platform authentication required"
          },
          "403": {
            "description": "Administrator role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Labs"
        ]
      }
    },
...
  },
  "info": {
    "title": "NGX coding-labs API",
    "description": "",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateLabDto": {
        "type": "object",
        "properties": {
          "workshopId": {
            "type": "string"
          },
          "workshopDocumentGroupId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tags": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "intro",
              "easy",
              "medium",
              "hard"
            ]
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "default": "draft"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        },
        "required": [
          "workshopId",
          "slug",
          "title",
          "createdBy"
        ]
      },
      "HandsOnLabMongo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "workshopId": {
            "type": "string"
          },
          "workshopDocumentGroupId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "intro",
              "easy",
              "medium",
              "hard"
            ]
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "currentDraftVersionId": {
            "type": "string"
          },
          "latestPublishedVersionId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string"
          },
          "archivedBy": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "workshopId",
          "slug",
          "title",
          "tags",
          "status",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy"
        ]
      },
      "UpdateLabDto": {
        "type": "object",
        "properties": {
          "workshopDocumentGroupId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tags": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "intro",
              "easy",
              "medium",
              "hard"
            ]
          },
          "estimatedMinutes": {
            "type": "number"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
...
    }
  }
}